+++ src/pages/gallery/GalleryDetailPage.tsx
@@ -1,0 +1,7 @@
+export default function GalleryDetailPage() {
+  return (
+    <div>
+      <p>갤러리페이지 공사 중..</p>
+    </div>
+  );
+}
